Add support for 64 bit mode

+++ libc/stdlib.c
@@ -1,6 +1,8 @@
 #include <stdlib.h>
 #include <stdint.h>
 #include "../libc/mem.h"
+#include "../drivers/serial.h"
+#include <debug.h>
 
 void exit() {
     /* Exectute an exit syscall */
@@ -9,6 +11,6 @@ void exit() {
                 int $0x80");
 }
 
-void malloc(uint32_t size) {
+void malloc(uint64_t size) {
     kmalloc(size);
 }
